// src/motorcycle-part/domain/enums/part-type.enum.ts
export enum PartType {
  FRONT_TIRE = 'FRONT_TIRE',
  REAR_TIRE = 'REAR_TIRE',
  FRONT_BRAKE_PADS = 'FRONT_BRAKE_PADS',
  REAR_BRAKE_PADS = 'REAR_BRAKE_PADS',
  FRONT_BRAKE_DISC = 'FRONT_BRAKE_DISC',
  REAR_BRAKE_DISC = 'REAR_BRAKE_DISC',
  BRAKE_FLUID = 'BRAKE_FLUID',
  CHAIN = 'CHAIN',
  FRONT_SPROCKET = 'FRONT_SPROCKET',
  REAR_SPROCKET = 'REAR_SPROCKET',
  ENGINE_OIL = 'ENGINE_OIL',
  OIL_FILTER = 'OIL_FILTER',
  AIR_FILTER = 'AIR_FILTER',
  SPARK_PLUGS = 'SPARK_PLUGS',
  COOLANT = 'COOLANT',
  FRONT_FORK_OIL = 'FRONT_FORK_OIL',
  REAR_SHOCK_OIL = 'REAR_SHOCK_OIL',
  BATTERY = 'BATTERY',
  CUSTOM = 'CUSTOM',
}